LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/template-shadowrootmode-img-src.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_not_equals: speculative case did not fetch got disallowed value ""

    <!-- speculative case in document.write -->
    <div><template shadowrootmode="closed"><img src="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=d8ca7a97-e8c3-4d61-8912-aff8350abd50&amp;encodingcheck=&Gbreve;"></template></div>


FAIL Speculative parsing, document.write(): template-shadowrootmode-img-src Unhandled rejection: assert_not_equals: speculative case did not fetch got disallowed value ""
